requirements.txt
setuptools>=68.0.0
wheel>=0.40.0
numpy<2.0.0
torch>=2.2.0
openai-whisper>=20231117
deep-translator>=1.11.4
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.7.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
tqdm>=4.66.1
librosa>=0.10.0
scipy>=1.11.0 